refactor(): use imported 'console'
