+++ dgram.js
@@ -11,6 +11,7 @@ import { EventEmitter } from './events.js'
 import { Buffer } from './buffer.js'
 import { isIPv4 } from './net.js'
 import * as ipc from './ipc.js'
+import console from './console.js'
 import dns from './dns.js'
 import gc from './gc.js'
