+++ fs/index.js
@@ -28,7 +28,8 @@ import { isBufferLike, isFunction } from '../util.js'
 import * as constants from './constants.js'
 import * as promises from './promises.js'
 import { Stats } from './stats.js'
-import * as ipc from '../ipc.js'
+import console from '../console.js'
+import ipc from '../ipc.js'
 import fds from './fds.js'
 import gc from '../gc.js'
