+++ index.js
@@ -3,6 +3,7 @@ export { Bluetooth } from './bluetooth.js'
 export { bootstrap } from './bootstrap.js'
 export * as buffer from './buffer.js'
 export { Buffer } from './buffer.js'
+export { default as console } from './console.js'
 export * as crypto from './crypto.js'
 export * as dgram from './dgram.js'
 export * as dns from './dns.js'
