+++ package.json
@@ -31,7 +31,8 @@
   "devDependencies": {
     "acorn": "8.8.0",
     "acorn-walk": "8.2.0",
-    "standard": "^17.0.0"
+    "standard": "^17.0.0",
+    "esbuild": "^0.16.16"
   },
   "optionalDependencies": {
     "@socketsupply/stream-relay": "^1.0.0"
